+++ sandbox/test_classic/src/proprocess/__init__.py
@@ -1,4 +1,2 @@
 
 
-
-from .config import SETTINGS, get_config